refactor: ts migration for poi/expired
